# Optimizations to the main layout loop