## What does the pull request do?
<!--- Give a bit of background on the PR here, together with links to with related issues etc. -->
When profiling my recent grid splitter PR I noticed that layout loop is doing a lot of unnecessary work (insane amount of property metadata calls, uncached delegates, etc.). This PR focuses on fixing low hanging fruit that brings benefits across entire application. 

I am using existing benchmark setup in `Avalonia.Benchmark` but I had to change few things to make it actually usable for consistent benchmarking (I will try to clean it up and submit it soon).

Benchmark creates around 6k controls and measures/arranges them all.

## What is the current behavior?
<!--- If the PR is a fix, describe the current incorrect behavior, otherwise delete this section. -->

|    Method |     Mean |     Error |    StdDev |    Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |---------:|----------:|----------:|---------:|------:|------:|----------:|
| Remeasure | 18.11 ms | 0.2199 ms | 0.1949 ms | 156.2500 |     - |     - | 725.62 KB |

## What is the updated/expected behavior with this PR?
<!--- Describe how to test the PR. -->

|    Method |     Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |---------:|----------:|----------:|------:|------:|------:|----------:|
| Remeasure | 10.63 ms | 0.2108 ms | 0.1972 ms |     - |     - |     - |      40 B |

## How was the solution implemented (if it's not obvious)?
<!--- Include any information that might be of use to a reviewer here. -->
I did split most of my work into separate commits to make the review easier. Notable parts are:
- removal of all LINQ from `LayoutQueue` plus keeping extra item buffer around so we don't have to allocate a new one per loop cycle.
- remove LINQ from visual children iteration in `Layoutable` (both arrange and measure overrides methods).
- remove a few `GetValue` calls from arrange and measure functions in `Layoutable`. This saves a lot of lookups as these functions are called very frequently and `GetValue` is quite expensive.
- Also I noticed that current use of `Contract.Requires` produces quite bad code so I decided to remove it from the hot path for now. Might need to be addressed again when we enable nullable reference types.
- In the end I wanted to check why `GetValue` is so expensive - `GetMetadata` call is a big factor due to metadata overrides and extra dictionary lookups caused by it. I noticed that most of the properties used in the layout do not have overrides so I implemented a fast path that does no dictionary lookups and caching if there are no overrides present.
- In the end all these small things combined lead to quite significant improvements (metadata change is the biggest factor).

## Checklist

- [ ] Added unit tests (if possible)?
- [ ] Added XML documentation to any related classes?
- [ ] Consider submitting a PR to https://github.com/AvaloniaUI/Avaloniaui.net with user documentation