@MarchingCube 2019-11-03
I just noticed that I made a wrong assumption about default metadata for direct properties. Will check how to fix it.