@kekekeks 2019-11-03
>Also I noticed that current use of Contract.Requires produces quite bad code so I decided to remove it from the hot path for now. 

Might be worth to have it under `DEBUG` conditional compilation symbol.