@MarchingCube 2019-11-03
Well, I think what https://github.com/AvaloniaUI/Avalonia/pull/3032 adds a valid replacement. With proper inlining it should be fine.